+++ components/browser/errorpages/src/main/java/mozilla/components/browser/errorpages/ErrorPages.kt
@@ -33,7 +33,7 @@ object ErrorPages {
             .use { it.readText() }
             .replace("%pageTitle%", context.getString(R.string.mozac_browser_errorpages_page_title))
             .replace("%backButton%", context.getString(R.string.mozac_browser_errorpages_page_go_back))
-            .replace("%button%", context.getString(R.string.mozac_browser_errorpages_page_refresh))
+            .replace("%button%", context.getString(errorType.buttonRes))
             .replace("%messageShort%", context.getString(errorType.titleRes))
             .replace("%messageLong%", context.getString(errorType.messageRes, uri))
             .replace("%css%", css)
@@ -45,7 +45,8 @@ object ErrorPages {
  */
 enum class ErrorType(
     @StringRes val titleRes: Int,
-    @StringRes val messageRes: Int
+    @StringRes val messageRes: Int,
+    @StringRes val buttonRes: Int = R.string.mozac_browser_errorpages_page_refresh
 ) {
     UNKNOWN(
         R.string.mozac_browser_errorpages_generic_title,
@@ -108,8 +109,9 @@ enum class ErrorType(
         R.string.mozac_browser_errorpages_invalid_content_encoding_message
     ),
     ERROR_UNKNOWN_HOST(
-        R.string.mozac_browser_errorpages_unknown_host_title,
-        R.string.mozac_browser_errorpages_unknown_host_message
+        R.string.mozac_browser_errorpages_unknown_host_title2,
+        R.string.mozac_browser_errorpages_unknown_host_message2,
+        R.string.mozac_browser_errorpages_unknown_host_button
     ),
     ERROR_MALFORMED_URI(
         R.string.mozac_browser_errorpages_malformed_uri_title,